+++ src/main/java/com/cburch/logisim/std/arith/floating/FpToInt.java
@@ -66,7 +66,7 @@ public class FpToInt extends InstanceFactory {
         new Object[] {BitWidth.create(8), BitWidth.create(32), ROUND_OPTION});
     setKeyConfigurator(new BitWidthConfigurator(StdAttr.FP_WIDTH));
     setOffsetBounds(Bounds.create(-40, -20, 40, 40));
-    setIcon(new ArithmeticIcon("FP\u2192I", 2));
+    setIcon(new ArithmeticIcon("FP→I", 2));
 
     final var ps = new Port[3];
     ps[IN] = new Port(-40, 0, Port.INPUT, StdAttr.FP_WIDTH);
@@ -84,7 +84,7 @@ public class FpToInt extends InstanceFactory {
     g.setColor(new Color(AppPreferences.COMPONENT_COLOR.get()));
     painter.drawBounds();
     painter.drawPort(IN);
-    painter.drawPort(OUT, "F\u2192I", Direction.WEST);
+    painter.drawPort(OUT, "F→I", Direction.WEST);
     painter.drawPort(ERR);
   }